LPS-107776 Move constants to the API module to be able to reuse them
